% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicaltrialr.R
\docType{package}
\name{clinicaltrialr-package}
\alias{clinicaltrialr}
\alias{clinicaltrialr-package}
\title{clinicaltrialr: Import study records from ClinicalTrials.gov}
\description{
It downloads the full study records of the desired study from
ClinicalTrials.gov and yields a dataset of the most commonly used fields.
}
\author{
\strong{Maintainer}: Stylianos Serghiou \email{stelios.serghiou@gmail.com}

}
