% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_read_results.R
\name{ct_read_results}
\alias{ct_read_results}
\title{Download the whole record of a single study}
\usage{
ct_read_results(URL)
}
\arguments{
\item{URL}{The URL given after searching on ClinicalTrials.gov.}
}
\value{
A dataframe of the ClinicalTrials.gov results table.
}
\description{
Returns the study record as an xml_document object.
}
\examples{
\dontrun{
# Copy-paste the URL from the ClinicalTrials.gov website
URL <- "http://www.clinicaltrials.gov/ct2/results?cond=Heart+Failure"
# Download results table
results_table <- get_results_table(URL)
}
}
