% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_read_trial_csv.R
\name{ct_read_trial_csv}
\alias{ct_read_trial_csv}
\title{Read commonly used information from ClinicalTrials.gov}
\usage{
ct_read_trial_csv(xml_document)
}
\arguments{
\item{xml_document}{An xml_document from `get_xml_document`}
}
\value{
A dataframe with the commonly used fields; if a field was not found,
    then this field will be empty, i.e. `character()`
}
\description{
Returns the commonly used information from ClinicalTrials.gov.
}
\examples{
\dontrun{
# Get the XML of study NCT03478891
xml_document <- ct_read_trial_xml("NCT03478891")
# Extract fields of interest
study <- extract_fields(xml_document)
}
}
