% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_read_trial_xml.R
\name{ct_read_trial_xml}
\alias{ct_read_trial_xml}
\title{Download the whole record of a single study}
\usage{
ct_read_trial_xml(NCT)
}
\arguments{
\item{NCT}{The NCT code of the desired study as a character,
e.g. "NCT03478891".}
}
\value{
The XML study record as an xml_document and xml_node.
}
\description{
Returns the study record as an xml_document object.
}
