% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-fields.R
\name{extract_fields}
\alias{extract_fields}
\title{Get commonly used information from ClinicalTrials.gov}
\usage{
extract_fields(xml_document)
}
\arguments{
\item{xml_document}{An xml_document from `get_xml_document`}
}
\value{
A dataframe with the commonly used fields; if a field was not found,
    then this field will be empty, i.e. `character()`
}
\description{
Returns the commonly used information from ClinicalTrials.gov.
}
\examples{
\dontrun{
# Get the XML of study NCT03478891
xml_document <- get_xml_document("NCT03478891")
# Extract fields of interest
study <- extract_fields(xml_document)
}
}
