% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-xml-document.R
\name{get_xml_document}
\alias{get_xml_document}
\title{Download the whole record of a single study}
\usage{
get_xml_document(NCT)
}
\arguments{
\item{NCT}{CharacterThe NCT code of the desired study as a character,
e.g. "NCT03478891".}
}
\value{
The XML study record as an xml_document and xml_node.
}
\description{
Returns the study record as an xml_document object.
}
