% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-xml-document.R
\name{write_xml_document}
\alias{write_xml_document}
\title{Save a downloaded xml_document as an XML file}
\usage{
write_xml_document(xml_document, path = "", ...)
}
\arguments{
\item{xml_document}{The xml_document received from `get_xml_document`.}

\item{path}{The path to the directory in which the file should be saved.}

\item{...}{Additional arguments to be passed to `xml2::write_xml`.}
}
\value{
The study record as an XML file in the designed location.
}
\description{
Saves the study record as an XML in the specified path.
}
\examples{
\dontrun{
# Get the XML of study NCT03478891
xml_document <- get_xml_document("NCT03478891")
# Save the XML
tmp <- tempfile()  # only required for this example, not in using the library
write_xml_document(xml_document, tmp)
}
}
