% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-fields.R
\name{.get_text}
\alias{.get_text}
\title{Get the desired text from the xml_document}
\usage{
.get_text(xml_document, xpath)
}
\arguments{
\item{xml_document}{An xml_document from `get_xml_document`}

\item{xpath}{The XPath as a character, e.g. "id_info/nct_id"}
}
\value{
The desired text as a character; if not found, then `character()`
}
\description{
Returns the text desired according to xpath.
}
